We modified the test_buttons.c source so that we can test all of the buttons.
I also added a linker script that should do the job of putting the various pieces (like boot_start.o and boot_loader.o) in the right place in memory.  I had to do this because we weren't guaranteed that the assembler would place main at the first address of a program's object code, which was causing dhrystone to crash badly.  The linker script is NOT yet part of the toolchain.